Revert "Merge branch 'master' into Gavin-01-27-20"
This reverts commit 7843327be3fcf712873460242b6bcb83aed87e3c, reversing
changes made to 94e57a91bd167515ff183f2eff3d5830cb034e17.
