+++ Assets/2D/MainMenuBG.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b4c3439a45200f546b15998dd7133bfe
+guid: 84a4109488052cf4299bd228290abe73
 TextureImporter:
   fileIDToRecycleName: {}
   externalObjects: {}
