+++ Assets/2D/ShotgunShells.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 55097767c9ecc6f428ff4fc69c8abeec
+guid: d5244a8cb0b35ab4f95eb439b0e7f8ec
 TextureImporter:
   fileIDToRecycleName: {}
   externalObjects: {}
