+++ Assets/Prefabs/AlienHead.0.1.blend.meta
@@ -149,7 +149,6 @@ ModelImporter:
     9500000: //RootNode
     10800000: Lamp
     10800002: Lamp.001
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
