+++ Assets/Prefabs/Cupboads_Sides_W_Sinks.2.blend.meta
@@ -59,7 +59,6 @@ ModelImporter:
     4300018: Table Top
     7400000: Default Take
     9500000: //RootNode
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
