+++ Assets/Prefabs/Cupboads_Sides_W_Sinks.blend.meta
@@ -172,7 +172,6 @@ ModelImporter:
     4300060: Table Top
     7400000: Default Take
     9500000: //RootNode
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
