+++ Assets/Prefabs/Enemy_Head.1.2.blend1.meta
@@ -1,9 +1,5 @@
 fileFormatVersion: 2
-<<<<<<< HEAD
-guid: 7c54da1d9db3f874dbf1e8b215d8fd49
-=======
 guid: e6105545043b1d146b27c5aa7ae2b139
->>>>>>> ae9d8abd6b2d9bec8060391ca34ca83abc546201
 DefaultImporter:
   externalObjects: {}
   userData:
