+++ Assets/Prefabs/Enemy_KneesN'Toes.0.blend.meta
@@ -12,7 +12,6 @@ ModelImporter:
     4300000: Cube
     7400000: Default Take
     9500000: //RootNode
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
