+++ Assets/Prefabs/Enemy_Shoulders.1.3.blend.meta
@@ -237,7 +237,6 @@ ModelImporter:
     13700032: Cylinder.002
     13700034: Torus.001
     13700036: Torus.002
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
