+++ Assets/Prefabs/PlusSymbol.blend.meta
@@ -21,7 +21,6 @@ ModelImporter:
     7400000: Default Take
     9500000: //RootNode
     10800000: Lamp
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
