+++ Assets/Prefabs/Quarantine.blend.meta
@@ -54,7 +54,6 @@ ModelImporter:
     4300016: Cylinder.001
     7400000: Default Take
     9500000: //RootNode
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
