+++ Assets/Prefabs/Sink.0.2.blend.meta
@@ -54,7 +54,6 @@ ModelImporter:
     4300016: Cube.001
     7400000: Default Take
     9500000: //RootNode
-    2186277476908879412: ImportLogs
   externalObjects: {}
   materials:
     importMaterials: 1
