+++ Assets/Prefabs/lureGrenade.prefab
@@ -19,7 +19,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &1355346988739301370
 Transform:
   m_ObjectHideFlags: 0
