+++ Assets/Scripts/AI scripts/enemyBase.cs
@@ -64,7 +64,6 @@ public class enemyBase : MonoBehaviour
             }
 
         }
-
     }
 
     // LateUpdate for physics
@@ -77,13 +76,10 @@ public class enemyBase : MonoBehaviour
             Debug.DrawRay(this.transform.position, direction, Color.green);     //for the intended path
 
             if (direction.magnitude > accuracy)                                 //If direction length is larger than enemy dis from player
-            {
-                this.transform.Translate(direction.normalized * speed * Time.deltaTime, Space.World);       //..Then move towards the player in global space
-                audioManager.PlaySound("RobotSound");
-            }
-        }                                                                                              
 
-
+                this.transform.Translate(direction.normalized * speed * Time.deltaTime, Space.World);       //..Then move towards the player in
+                                                                                                            //in global space
+        }
     }
 
     //void FixedUpdate()
