+++ Assets/Scripts/Player/Player.cs
@@ -31,7 +31,7 @@ public class Player : MonoBehaviour
         characterController = GetComponent<CharacterController>();
         gunEquipper = GetComponent<GunEquipper>();
         healthBar.setMaxHealth(maxHealth);
-        //currency = GetComponent<GameManager>().bubblegum;
+        currency = GetComponent<GameManager>().bubblegum;
     }
 
     //added takeDamage function
@@ -39,7 +39,7 @@ public class Player : MonoBehaviour
     {
         int healthDamage = 1;
         curHealth -= healthDamage;
-        //healthBar.setHealth(curHealth);
+        healthBar.setHealth(curHealth);
         Debug.Log("you've been hurt, health is: " + curHealth + " out of: " + maxHealth);
         if(curHealth <= 0)
         {
