+++ Assets/Scripts/gameplay & world/GunEquipper.cs.meta
@@ -3,13 +3,7 @@ guid: bd98973668d7df746bb8da0398f98978
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences:
-  - pistol: {instanceID: 0}
-  - shotgun: {fileID: 179884741893322542, guid: ebeca8ea42470c044a92dfa9c7d2d6bc,
-      type: 3}
-  - fragGrenade: {instanceID: 0}
-  - lureGrenade: {instanceID: 0}
-  - miniGun: {instanceID: 0}
+  defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
